style: annotations renaming
