+++ service-member/module-api/src/main/java/com/github/msafriends/moduleapi/validator/MemberNameValidator.java
@@ -1,6 +1,6 @@
 package com.github.msafriends.moduleapi.validator;
 
-import com.github.msafriends.moduleapi.validator.anntations.ValidMemberName;
+import com.github.msafriends.moduleapi.validator.annotations.ValidMemberName;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
