+++ service-member/module-api/src/main/java/com/github/msafriends/moduleapi/validator/PhoneNumberValidator.java
@@ -1,6 +1,6 @@
 package com.github.msafriends.moduleapi.validator;
 
-import com.github.msafriends.moduleapi.validator.anntations.ValidPhoneNumber;
+import com.github.msafriends.moduleapi.validator.annotations.ValidPhoneNumber;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
