+++ service-member/module-api/src/main/java/com/github/msafriends/moduleapi/validator/annotations/ValidPassword.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.moduleapi.validator.anntations;
+package com.github.msafriends.moduleapi.validator.annotations;
 
 import com.github.msafriends.moduleapi.validator.PasswordValidator;
 import jakarta.validation.Constraint;
